pyproject.toml
[tool.poetry]
name = "ubex_autapi"
version = "1.0.0"
description = "Test task for the Ubex team"
authors = ["Nick Sablukov <dessanndes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
django-configurations = "^2.2"
dj_database_url = "^0.5.0"
psycopg2 = "^2.8.5"
attrs = "^19.3.0"
djangorestframework = "^3.11.0"
drf_yasg = "^1.17.1"
django-filter = "^2.3.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
mypy = "^0.781"
flake8 = "^3.8.3"
bandit = "^1.6.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
